requirements.txt
streamlit>=1.42.0
langgraph>=1.1.0
langchain-core>=1.2.0
langchain-openai>=0.1.0
openai>=1.0.0
pymupdf>=1.27.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
pydantic>=2.0.0
